deleteFileTree now has exclude, and allowing keychain.txt and ro_cache.txt to be excluded.
svn path=/dasCore/community/autoplot2011/trunk/; revision=6801

+++ dasCoreUtil/src/org/das2/util/FileUtil.java
@@ -6,6 +6,12 @@
 package org.das2.util;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.nio.channels.FileChannel;
+import java.util.Set;
 
 /**
  * static utility methods.
@@ -21,19 +27,37 @@ public class FileUtil {
     
     /**
      * deletes all files and folders below root, and root, just as "rm -r" would.
-     * TODO: check links
-     * @throws IllegalArgumentException if it is unable to delete a file
+     * TODO: check links     
+     * @param root the root where we start deleting.
      * @return true if the operation was successful.
+     * @throws IllegalArgumentException if it is unable to delete a file
      */
     public static boolean deleteFileTree(File root) throws IllegalArgumentException {
+        return deleteFileTree( root, null );
+    }
+
+    /**
+     * deletes all files and folders below root, and root, just as "rm -r" would, excluding
+     * any files named in exclude.  For example, exclude could contain "readme.txt".
+     * @param root the root where we start deleting.
+     * @param exclude null or a set containing names to exclude.
+     * @return true if the operation was successful.
+     * @throws IllegalArgumentException if it is unable to delete a file
+     */
+    public static boolean deleteFileTree( File root, Set<String> exclude ) throws IllegalArgumentException {
         if (!root.exists()) {
             return true;
         }
         File[] children = root.listFiles();
         boolean success = true;
+        boolean noExclude= true;
         for (int i = 0; i < children.length; i++) {
+            if ( exclude!=null && exclude.contains(children[i].getName()) ) {
+                noExclude= false;
+                continue;
+            }
             if (children[i].isDirectory()) {
-                success = success && deleteFileTree(children[i]);
+                success = success && deleteFileTree(children[i],exclude);
             } else {
                 success = success && ( !children[i].exists() || children[i].delete() ); // in case file is deleted by another process, check exists again.
                 if (!success) {
@@ -41,16 +65,18 @@ public class FileUtil {
                 }
             }
         }
-        success = success && (!root.exists() || root.delete());
+        if ( noExclude ) {
+            success = success && (!root.exists() || root.delete());
+        }
         if (!success) {
             throw new IllegalArgumentException("unable to delete folder " + root);
         }
-        return success;
+        return success;        
     }
-
+    
     /**
      * deletes all files with the given name, and root, just as "find , -name name -exec rm {} \;" would.
-     * TODO: check links
+     * TODO: check links.  For example deleteWithinFileTree( root, ".listing" )
      * @throws IllegalArgumentException if it is unable to delete a file
      * @return true if the operation was successful.
      */
@@ -73,5 +99,41 @@ public class FileUtil {
             }
         }
         return success;
+    }
+    
+    /**
+     * copies the file or folder from src to dst.
+     * @param src the source file or folder.
+     * @param dst the location for the new file or folder.
+     * @throws FileNotFoundException
+     * @throws IOException 
+     */
+    public static void fileCopy( File src, File dst ) throws FileNotFoundException, IOException {
+        if ( src.isDirectory() && ( !dst.exists() || dst.isDirectory() ) ) {
+            if ( !dst.exists() ) {
+                if ( !dst.mkdirs() ) throw new IOException("unable to mkdir " + dst);
+            }
+            File dst1= new File( dst, src.getName() );
+            if ( !dst1.exists() && !dst1.mkdir() ) throw new IOException("unable to mkdir " + dst1);
+            dst= dst1;
+            File[] files= src.listFiles();
+            for ( File f:files ) {
+                if ( f.isDirectory() ) {
+                    dst1= new File( dst, f.getName() );
+                    if ( !dst1.exists() && !dst1.mkdir() ) throw new IOException("unable to mkdir " + dst1);
+                } else {
+                    dst1= dst;
+                }
+                fileCopy( f, dst1 );
+            }
+            return;
+        } else if ( dst.isDirectory() ) {
+            dst= new File( dst, src.getName() );
+        }
+        FileChannel ic = new FileInputStream(src).getChannel();
+        FileChannel oc = new FileOutputStream(dst).getChannel();
+        ic.transferTo(0, ic.size(), oc);
+        ic.close();
+        oc.close();
     }
 }
